requirements.txt
asgiref==3.4.1
cffi==1.14.6
codegen==1.0
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.1.3
django-extensions==3.1.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
psycopg2==2.9.1
pycparser==2.20
pydot==1.4.2
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
Werkzeug==2.0.1
whitenoise==5.3.0
